refactor: CardCompanyContext 모듈을 domain/context 폴더 하위로 이동
